renamed 'base' to 'core'. Refactored
