--- torchstain/base/normalizers/__init__.py
@@ -1,1 +1,0 @@
-import torchstain.base.normalizers.macenko_normalizer
