+++ torchstain/core/normalizers/__init__.py
@@ -1,0 +1,1 @@
+import torchstain.core.normalizers.macenko_normalizer
