+++ torchstain/tf_backend/normalizers/tensorflow_macenko_normalizer.py
@@ -1,5 +1,5 @@
 import tensorflow as tf
-from torchstain.base.normalizers.he_normalizer import HENormalizer
+from torchstain.core.normalizers.he_normalizer import HENormalizer
 from torchstain.tf_backend.utils import cov_tf, percentile_tf, solveLS
 import numpy as np
 import tensorflow.keras.backend as K
